pyproject.toml
[project]
name = "ftl-translator"
version = "0.1.0"
description = ""
authors = [{ name = "taimast", email = "charlskenno@gmail.com" }]
requires-python = ">=3.10,<4.0"
readme = "README.md"
dependencies = [
    "fluent-compiler>=0.3",
    "deep-translator>=1.11.4,<2",
    "fluent-runtime>=0.3",
    "aiohttp[speedups]>=3.9",
    "aiofiles>=20.0.0",
    "openai>=1.55.1,<2",
    "aiohttp-socks>=0.10.1,<0.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"
